Split module declarations into seperate files
