+++ src/Types/css.d.ts
@@ -1,0 +1,1 @@
+declare module "*.css";
