+++ src/Types/html.d.ts
@@ -1,0 +1,4 @@
+declare module "*.html" {
+    const value: string;
+    export default value;
+}
