+++ src/Types/txt.d.ts
@@ -1,0 +1,4 @@
+declare module "*.txt" {
+    const content: string;
+    export default content;
+}
